Fix disconnection when running demo docker compose
Rails error :
`exec': PQconsumeInput() server closed the connection unexpectedly (PG::ConnectionBad)

PG error:
server process (PID 365) was terminated by signal 13: Broken pipe

After analysis, it's the health check that fails when running frequent
queries on database during the benchmark. As it fails, benchmark is killed.
And then PG is stopped.

To avoid that we get rid of the query to check that data have been written.
We rather rely on the creation of file once all init migrations are done.

Notes:
Increasing timeout from 2 to 200s to support slow queries is not enough

Setting jit=off on database, but does not help
https://stackoverflow.com/questions/63768188/postgresql-pg-hba-conf-postgresql-conf-the-connection-to-the-server-was-lost

According to docker stats, postgres server uses approximately up to 300MiB